# apps/web/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SystemOneStudio Control Plane",
    "description": "Control plane for the zero-code decision model training platform. The prediction surface is served separately per deployment at /v1/predict.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/scenes": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "已创建场景",
        "operationId": "list_scenes_api_scenes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SceneSummary"
                  },
                  "type": "array",
                  "title": "Response List Scenes Api Scenes Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "从模板或从零创建场景",
        "description": "Instantiate a template, or build a scene from scratch.\n\nCreating from scratch requires a name; instantiating a template inherits its\nrecommended model, methods and metrics.",
        "operationId": "create_scene_api_scenes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSceneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenes/{scene_id}/jev": {
      "patch": {
        "tags": [
          "platform"
        ],
        "summary": "更新 JEV 开关",
        "description": "The two switches stay independent (需求方案.txt 6.3).",
        "operationId": "update_jev_flags_api_scenes__scene_id__jev_patch",
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scene Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JevFlags"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "项目列表",
        "operationId": "list_projects_api_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  },
                  "type": "array",
                  "title": "Response List Projects Api Projects Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "新建项目",
        "operationId": "create_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/system/audit": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "审计日志",
        "description": "变更记录（需求方案.txt 原则 5 可审计）.\n\n``limit`` is bounded: an unbounded or negative slice returned the *last* N\nentries for a negative limit, which reads as the newest ones.",
        "operationId": "audit_log_api_system_audit_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEntry"
                  },
                  "title": "Response Audit Log Api System Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditEntry": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "ts": {
            "type": "string",
            "title": "Ts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "actor",
          "action",
          "target",
          "detail",
          "ts"
        ],
        "title": "AuditEntry"
      },
...
      "CreateProjectRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "scene_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Id"
          },
          "mode": {
            "$ref": "#/components/schemas/UiMode",
            "default": "wizard"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateProjectRequest"
      },
      "CreateSceneRequest": {
        "properties": {
          "template_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldSpec"
            },
            "type": "array",
            "title": "Fields"
          },
          "labels": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array",
            "title": "Labels"
          },
          "jev_format_compat": {
            "type": "boolean",
            "title": "Jev Format Compat",
            "default": true
          },
          "jev_training_compat": {
            "type": "boolean",
            "title": "Jev Training Compat",
            "default": true
          }
        },
        "type": "object",
        "title": "CreateSceneRequest"
      },
      "Decision": {
        "type": "string",
        "enum": [
          "black",
          "white",
          "gray"
        ],
        "title": "Decision",
        "description": "Three-way judgement outcome.\n\nDeliberately NOT a boolean. The `gray` label is a first-class outcome that\nthreads through synthesis ratios, quality reports and per-label evaluation\nmetrics. Collapsing it into black/white silently drops the hardest cases."
      },
...
      "FieldSpec": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "kind": {
            "type": "string",
            "pattern": "^(numeric|categorical|text|datetime)$",
            "title": "Kind",
            "default": "numeric"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": false
          },
          "sensitive": {
            "type": "boolean",
            "title": "Sensitive",
            "default": false
          },
          "constraint_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraint Min"
          },
          "constraint_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraint Max"
          }
        },
        "type": "object",
        "required": [
          "name",
          "label"
        ],
        "title": "FieldSpec",
        "description": "One field in a scene's schema."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JevFlags": {
        "properties": {
          "jev_format_compat": {
            "type": "boolean",
            "title": "Jev Format Compat",
            "default": true
          },
          "jev_training_compat": {
            "type": "boolean",
            "title": "Jev Training Compat",
            "default": true
          }
        },
        "type": "object",
        "title": "JevFlags",
        "description": "The two JEV compatibility switches.\n\nThese are two independent switches by requirement, not one toggle. They\ncontrol different layers and must be allowed to differ:\n\n* ``jev_format_compat``   -> input/output schema, prompt format, API shape\n* ``jev_training_compat`` -> training paradigm, hyperparameters, eval basis\n\nBoth default to enabled. When ``jev_format_compat`` is off the API still\nreturns the same four fields so callers keep a single parser, but the\nresponse is explicitly flagged as non-JEV via ``PredictResponse``'s\n``jev_compatible`` field."
      },
...
      "Project": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "mode": {
            "$ref": "#/components/schemas/UiMode"
          },
          "scene_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Code"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "mode",
          "scene_code",
          "created_at"
        ],
        "title": "Project"
      },
...
      "SceneSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "field_count": {
            "type": "integer",
            "title": "Field Count"
          },
          "labels": {
            "items": {
              "$ref": "#/components/schemas/Decision"
            },
            "type": "array",
            "title": "Labels"
          },
          "recommended_base_model": {
            "type": "string",
            "title": "Recommended Base Model"
          },
          "recommended_methods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Methods"
          },
          "core_metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Core Metrics"
          },
          "jev_format_compat": {
            "type": "boolean",
            "title": "Jev Format Compat"
          },
          "jev_training_compat": {
            "type": "boolean",
            "title": "Jev Training Compat"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "template | custom"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name",
          "summary",
          "field_count",
          "labels",
          "recommended_base_model",
          "recommended_methods",
          "core_metrics",
          "jev_format_compat",
          "jev_training_compat",
          "source"
        ],
        "title": "SceneSummary"
      },
...
      "UiMode": {
        "type": "string",
        "enum": [
          "wizard",
          "canvas",
          "rapid",
          "expert"
        ],
        "title": "UiMode",
        "description": "Interaction modes. All four share one engine and one flow definition."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}